#20 wip

+++ AdaTranslation.Application/DTOs/ServiceDto.cs
@@ -1,7 +1,4 @@
 ﻿namespace AdaTranslation.Application.DTOs
 {
-    public record ServiceDto(
-        long Id, 
-        string Description
-        );
+    public record ServiceDto(int Id, string Description);
 }
